Sorting of the rules
this patch allows sorting of the rules ... so user can click a column and sort it table by the column
it also introduces the default sorting rule which groups the rules by  category and is alphabetical withing category. 
If you create multiple clones of the same rule, you will see them in the list sorted by time.